Rename "materials" package back to "material"
There are still some compatibility issues, but not as many.
