+++ hexrd/__init__.py
@@ -1,8 +1,7 @@
-from .materials import symbols
-from .materials import symmetry
-from .materials import spacegroup
-from .materials import crystallography
-from .materials import unitcell
-from .materials import material
+from .material import crystallography
+from .material import jcpds
 from .material import mksupport
-from .materials import jcpds
+from .material import spacegroup
+from .material import symbols
+from .material import symmetry
+from .material import unitcell
