+++ hexrd/material/__init__.py
@@ -1,0 +1,7 @@
+from .material import (
+    _angstroms,
+    _kev,
+    load_materials_hdf5,
+    Material,
+    save_materials_hdf5,
+)
